third_party/blink/web_tests/flag-specific/enable-blink-features=ForceSynchronousHTMLParsing/virtual/text-antialias/custom-font-data-crash2-expected.txt

CONSOLE WARNING: line 29: Failed to decode downloaded font: custom-font-data-crash2.html
CONSOLE WARNING: line 29: OTS parsing error: invalid sfntVersion: 1008813135
PASS
